requirements.txt
google-genai==1.46.0
easygui==0.98.3
pygame-gui==0.6.14
pillow==10.4.0
